refactor(chatbot): use same logic for support requests as standard requests
